requirements.txt
# --- 核心 Web 与基础库 ---
fastapi==0.115.5
uvicorn==0.32.1
pydantic==2.12.5
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.3
aiofiles==25.1.0
numpy==1.26.4
pandas==2.3.0
Pillow==11.2.1
tqdm==4.67.1
nest-asyncio==1.6.0

# --- RAG 与 LLM ---
langchain==0.3.12
langchain-core==0.3.27
langchain-community==0.3.12
langchain-openai==0.2.9
openai==1.109.1
FlagEmbedding==1.3.5
rank-bm25==0.2.2
jieba==0.42.1
tiktoken==0.8.0
onnxruntime==1.22.0

# --- LlamaIndex (Excel/CSV 深度解析) ---
llama-index==0.14.13
llama-parse==0.6.54
llama-index-llms-deepseek==0.2.2
llama-index-embeddings-huggingface==0.6.1
llama-index-vector-stores-chroma==0.5.5
chromadb==1.4.1

# --- 文档解析 ---
pdfplumber==0.11.8
pypdf==6.2.0
python-docx==1.2.0
pypandoc==1.16
pdf2image==1.17.0

# --- 解析引擎与 OCR (DeepDoc/RAGFlow) ---
paddleocr==3.3.1
paddlepaddle==3.2.1
shapely==2.1.2
pyclipper==1.3.0.post6
opencv-python==4.11.0.86
roman-numbers==1.0.2
word2number==1.1
huggingface-hub==0.36.0
